Unify cli operators sturcture
