+++ packages/cli/src/@internal/operators/configure/index.ts
@@ -1,6 +1,5 @@
 import { Configure } from '../../utils'
-import { EtlxOptions } from '../../types'
-import { ConfigurationOptions } from './types'
+import { EtlxOptions, ConfigurationOptions } from '../../types'
 
 export const configure = (...xs: Configure<ConfigurationOptions>[]) => (opts: EtlxOptions): EtlxOptions => ({
     ...opts,
