+++ packages/cli/src/@internal/operators/configure/operators/addFile.ts
@@ -1,0 +1,6 @@
+import { ConfigurationOptions } from '../../../types'
+
+export const addFile = (filepath: string) => (opts: ConfigurationOptions) => ({
+    ...opts,
+    paths: [...opts.paths, filepath],
+})
