+++ packages/cli/src/@internal/operators/configure/operators/addObject.ts
@@ -1,0 +1,6 @@
+import { ConfigurationOptions } from '../../../types'
+
+export const addObject = (config: any) => (opts: ConfigurationOptions) => ({
+    ...opts,
+    objects: [...opts.objects, config],
+})
