+++ packages/cli/src/@internal/operators/configure/operators/addParser.ts
@@ -1,0 +1,6 @@
+import { FileParser, ConfigurationOptions } from '../../../types'
+
+export const addParser = (parser: FileParser) => (opts: ConfigurationOptions) => ({
+    ...opts,
+    parsers: [...opts.parsers, parser],
+})
