+++ packages/cli/src/@internal/operators/configure/operators/addSchema.ts
@@ -1,0 +1,6 @@
+import { Schema, ConfigurationOptions } from '../../../types'
+
+export const addSchema = (schema: Schema) => (opts: ConfigurationOptions) => ({
+    ...opts,
+    schemes: [...opts.schemes, schema],
+})
