+++ packages/cli/src/@internal/operators/configure/operators/createSchema.ts
@@ -1,0 +1,5 @@
+import { Schema } from '../../../types'
+
+export function createSchema<T>(schema: Schema<T>): Schema<T> {
+    return schema as any
+}
