+++ packages/cli/src/@internal/operators/configure/operators/index.ts
@@ -1,0 +1,6 @@
+export * from './createSchema'
+export * from './addSchema'
+export * from './addFile'
+export * from './addParser'
+export * from './addObject'
+export * from './warnings'
