+++ packages/cli/src/@internal/operators/configure/operators/warnings.ts
@@ -1,0 +1,6 @@
+import { ConfigurationOptions } from '../../../types'
+
+export const warnings = (enabled: boolean) => (opts: ConfigurationOptions) => ({
+    ...opts,
+    suppressWarnings: !enabled,
+})
